pyproject.toml
[project]
name = "tech-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "awswrangler>=3.14.0",
    "boto3>=1.42.17",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.2.1",
    "yfinance>=1.0",
]
